SD/requirements.txt
# ============================================
# EV CHARGING SYSTEM - Requirements
# ============================================
# Python 3.11+ required
# Install with: pip install -r requirements.txt
# ============================================

# Kafka Client
kafka-python==2.0.2

# WebSocket Server (para interfaces web)
websockets==12.0
aiohttp==3.9.1
aiosignal==1.3.1
frozenlist==1.4.0
multidict==6.0.4
yarl==1.9.2
attrs==23.1.0

# Async support
async-timeout==4.0.3

# Optional: for better terminal colors
colorama==0.4.6

# ============================================
# BUILT-IN (No need to install):
# - sqlite3 (database)
# - asyncio (async programming)
# - socket (TCP communication)
# - threading (concurrency)
# - json (data serialization)
# - hashlib (password hashing)
# ============================================
